backend/app_instance.py
from flask import Flask
from flask_cors import CORS
from flask_socketio import SocketIO
import os
from dotenv import load_dotenv

load_dotenv()
secret_key = os.getenv('SECRET_KEY')

app = Flask(__name__)
app.config['SECRET_KEY'] = secret_key
CORS(app)

socketio = SocketIO(app, cors_allowed_origins="*", async_mode="gevent")
