backend/requirements.txt
Flask==3.1.1
Werkzeug>=3.1.3
Flask_Cors==6.0.1
Flask_Login==0.6.3
progressbar2==4.5.0
psutil==7.0.0
PyJWT==2.10.1
python-dotenv==1.1.1
toml==0.10.2
requests==2.32.4
pick==2.4.0
install-jdk==1.1.0
beautifulsoup4==4.13.4
colorama
pyTelegramBotAPI
asyncio
aiohttp
flask-socketio
gevent